requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
decorator==5.1.1
discord==2.3.1
discord.py==2.3.1
Django==4.2.3
django-admin==2.0.2
django-excel-response2==3.0.6
django-filter==23.2
django-mercadopago==6.0.1
django-six==1.0.5
django-steamauth==1.1.2.3
djangorestframework==3.14.0
docopt==0.6.2
excel-base==1.0.4
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
isoweek==1.3.3
mercadopago==0.3.5
multidict==6.0.4
numpy==1.25.1
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pipreqs==0.4.13
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
screen==1.0.1
setuptools-git==1.2
setuptools-scm==7.1.0
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
validators==0.20.0
xlwt==1.3.0
yarg==0.1.9
yarl==1.9.2
